Separate app from main
